gotcha_flask/requirement.txt
flask==2.1.2
flask_sqlalchemy==2.5.1
joblib==1.1.0
pandas==1.4.2
numpy==1.22.3
google-cloud-bigtable==2.10.0
google-cloud-core==2.3.0
datetime==4.4
sklearn